coda: add CODA960 JPEG support
This patch adds JPEG encoding and decoding support for CODA960,
handling the JPEG hardware directly. Separate JPEG encoder and
decoder video devices are created due to the separate hardware
unit and different supported pixel formats.
The hardware can not change subsampling on the fly, but encode
and decode 4:2:2 subsampled JPEG images from and into this
format.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>