gpu: ipu-v3: ipu-ic: Add task state debugging and single shot trigger
The IPU documentation tells us that task disabling is not instantaneous,
but active tasks are allowed to finish. Immediately issuing a task disable
after the task becomes active should disable the task after processing of
a single frame has finished.
Judging from the task state register, this only works properly for the
post processing task, but not for the viewfinder task.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>