BACKPORT: ARM: imx6: remove PGC handling from GPC driver
This removes all functionality regarding the power gating controller
from the GPC driver. This will be added back in a re-written driver
outside of the architecture code.

This keeps only the IRQ controller code in the architecture, as this
is closely coupled to the CPU idle implementation.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>

Conflicts:
	arch/arm/mach-imx/gpc.c

Fixed conflict because dynamic state allocation in genpd had not been
cherry-picked (commit 59d65b73a23c).

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>