BACKPORT: drm/etnaviv: return GPU fence through the submit structure
The next patch will need the complete dma_fence, instead of just the seqno,
to create the sync_file in etnaviv_ioctl_gem_submit, in case an
out_fence_fd is requested.

The submit needs to hold a reference to the dma_fence, to avoid raceing
with the GPU completing the fence.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>

Conflicts:
	drivers/gpu/drm/etnaviv/etnaviv_gem.h
	drivers/gpu/drm/etnaviv/etnaviv_gem_submit.c
	drivers/gpu/drm/etnaviv/etnaviv_gpu.c

Fixed conflict given that the 'fence' struct and APIs aren't named
'dma_fence' (didn't cherry-pick commit f54d1867005c).

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>