BACKPORT: drm/etnaviv: don't put fence in case of submit failure
If we bail out of the submit before actually adding the cmdstream
to the kernel ring there is no valid fence to put. Make sure to skip
the fence_put in that case.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>

Conflicts:
	drivers/gpu/drm/etnaviv/etnaviv_gem_submit.c

Fixed conflict given that the 'fence' struct and APIs aren't named
'dma_fence' (didn't cherry-pick commit f54d1867005c).

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>